api/requirements.txt
fastapi[standard]
requests
sqlmodel
psycopg2-binary
pyjwt
python-dotenv